backend/requirements.txt
flask
flask-cors
onnxruntime
Pillow
numpy
gdown